requirements.txt
Flask==2.0.1
Flask-SQLAlchemy
Flask-Migrate
python-dotenv